requirements.txt
kubernetes~=24.2.0
kafka-python
argparse~=1.4.0